roachtest: use dmsetup disk-staller for operation
The cgroup-based disk staller has some weird interactions with
nvme ssds on GCP, where the cockroach process can become unkillable
if it is waiting on some disk operations on a disk that is stalled
using cgroup settings. This change updates the disk-stall operation
to use the dmsetup-based disk staller, which should hopefully be more
stable under this mode of operation.

Epic: none

Release note: None
